Add --usage-only and --usage-file options to gbagbe
A summary of reference usage and keepage is now written to stderr, this
necessitated cleaning up a bit of the internal mechanism.  The new
option --usage-only stops after pass 1 to provide this summary
information and exit without producing an output BAM.  Per-reference
statistics are dumped when given the name of a --usage-file.